requirements.txt
numpy==1.22.2
opencv-python==4.5.5.62
Pillow==9.0.1
google-cloud-vision==2.7.0